+++ CHANGELOG.md
@@ -34,6 +34,5 @@ Headings should be (only add headings as needed):
 - Fixed section heading closing tag in `content-espresso_events-header.php` ([541](https://github.com/eventespresso/event-espresso-core/pull/541))
 - Fixed typo in paypal pro ([535](https://github.com/eventespresso/event-espresso-core/pull/535))
 - Fixed EE CPT objects not being added to WP_Post objects during Ajax Requests ([531](https://github.com/eventespresso/event-espresso-core/pull/531))
+- Fixed transaction not being assigned to payment log entry when created using an EE_Transaction object ([545](https://github.com/eventespresso/event-espresso-core/pull/545))
 - Fixed shortcode and template tag loading during AJAX requests ([537](https://github.com/eventespresso/event-espresso-core/pull/537))
-
-